+++ examples/medplum-photon-integration/src/components/PatientPrescription.tsx
@@ -1,17 +1,16 @@
 import { Button, Group, Title } from '@mantine/core';
 import { notifications } from '@mantine/notifications';
-import { normalizeErrorString, PatchOperation } from '@medplum/core';
-import { Identifier, Patient } from '@medplum/fhirtypes';
+import { normalizeErrorString } from '@medplum/core';
+import { Patient } from '@medplum/fhirtypes';
 import { Document, useMedplum } from '@medplum/react';
 import { IconCircleCheck, IconCircleOff } from '@tabler/icons-react';
 import { useState } from 'react';
 
 interface PatientPrescriptionProps {
   patient: Patient;
-  onPatientChange: (patient: Patient) => void;
 }
 
-export function PatientPrescription({ patient, onPatientChange }: PatientPrescriptionProps): JSX.Element {
+export function PatientPrescription({ patient }: PatientPrescriptionProps): JSX.Element {
   const medplum = useMedplum();
 
   const patientSynced = patient.identifier?.find((id) => id.system === 'https://neutron.health/patients');
@@ -36,7 +35,7 @@ export function PatientPrescription({ patient, onPatientChange }: PatientPrescri
 
   async function syncPatient(): Promise<void> {
     try {
-      const photonPatientId = await medplum.executeBot(
+      await medplum.executeBot(
         {
           system: 'https://neutron.health/bots',
           value: 'sync-patient',
@@ -46,7 +45,6 @@ export function PatientPrescription({ patient, onPatientChange }: PatientPrescri
         },
         'application/fhir+json'
       );
-      await updatePatient(patient, photonPatientId);
       notifications.show({
         icon: <IconCircleCheck />,
         title: 'Success',
@@ -60,30 +58,6 @@ export function PatientPrescription({ patient, onPatientChange }: PatientPrescri
         title: 'Error',
         message: normalizeErrorString(err),
       });
-    }
-  }
-
-  async function updatePatient(patient: Patient, photonId: string): Promise<void> {
-    const identifiers = patient.identifier ?? [];
-    const photonIdentifier: Identifier = {
-      system: 'https://neutron.health/patients',
-      value: photonId,
-    };
-    identifiers.push(photonIdentifier);
-
-    const patientId = patient.id as string;
-
-    const op = patient.identifier ? 'replace' : 'add';
-    const ops: PatchOperation[] = [
-      { op: 'test', path: '/meta/versionId', value: patient.meta?.versionId },
-      { op, path: '/identifier', value: identifiers },
-    ];
-
-    try {
-      const updatedPatient = await medplum.patchResource('Patient', patientId, ops);
-      onPatientChange(updatedPatient);
-    } catch (err) {
-      console.error(err);
     }
   }
